Refactoring file processing
